# prefetch scale/bias