Summary:
When block_size is a multiple of cache line size, we also need to prefetch for scale and bias.
Thanks Wu Yong and Dhiraj Kalamkar for this suggestion!

Other minor fixes:
* Remove extra prefetches of 2/4-bit cases
* Fix a bug cache flushed case performance was not printed from EmbeddingSpMDM8BitBenchmark

Differential Revision: D20582936